+++ x/btcstkconsumer/types/events.go
@@ -1,7 +1,15 @@
 package types
 
-func NewConsumerRegisteredEvent(consumerId, consumerName, consumerDescription string, consumerType ConsumerType,
-	rollupFinalityContractAddress string) *EventConsumerRegistered {
+import (
+	"cosmossdk.io/math"
+)
+
+func NewConsumerRegisteredEvent(
+	consumerId, consumerName, consumerDescription string,
+	consumerType ConsumerType,
+	rollupFinalityContractAddress string,
+	babylonRewardsCommission math.LegacyDec,
+) *EventConsumerRegistered {
 	return &EventConsumerRegistered{
 		ConsumerId:          consumerId,
 		ConsumerName:        consumerName,
@@ -10,5 +18,6 @@ func NewConsumerRegisteredEvent(consumerId, consumerName, consumerDescription st
 		RollupConsumerMetadata: &RollupConsumerMetadata{
 			FinalityContractAddress: rollupFinalityContractAddress,
 		},
+		BabylonRewardsCommission: babylonRewardsCommission,
 	}
 }